Large page allocating memory allocators
The memory allocators have been expanded to handle allocating large pages.
Currently, we use the same pool of physical frames as the regular allocator,
so when we request for a large page after a few regular pages; the allocator
will skip the remaining regular pages and fast forward the pointer to the next
large page boundary to keep things simple.

Cons: physical memory frames will be wasted!

Change-Id: Ic781ff3e5a34cb61c7f8f120af02dabb251e408e

Refactoring to reuse code

Change-Id: I8c3f8a23482cca19d96b892526d34758ed17404f