cpu: add perfect Branch Predictor
This commit adds a oracle branch predictor based on a trace file.
This BP have 0 misprediction on committed branches.

The trace file is generated by an earlier run and contains in order
the pc, upc and target of each taken branch. An option was added to
generate this file with the simple and o3 cpu.

The perfect branch predictor works for now with only 1 cpu, no smt
and checkpoints are not supported.

The trace can be generated with '--trace-BP' on the simple or o3 CPU.
The trace can be used with '--perfect-BP'.

Change-Id: Ib5c9e61c26d12edc040de2e2751c62835465cb95