arch: Add instruction size and PC set methods
Add the instruction size of a static instruction. x86 and arm
decoders add the instruction size to the macro instruction.
However, microops are still handled by the fetch stage which
is not nice.
Furthermore, we add a set method to the PC state. It allows
setting a PC state to acertain address.
Both methods are required for the decoupled front-end.

Change-Id: I311fe3f637e867c42dee7781f5373ea2e69e2072
Signed-off-by: David Schall <david.schall@ed.ac.uk>